refactor schema metadata
