+++ sharding-core/sharding-core-common/src/main/java/org/apache/shardingsphere/core/metadata/ShardingMetaDataLoader.java
@@ -67,21 +67,35 @@ public final class ShardingMetaDataLoader {
     private final int maxConnectionsSizePerQuery;
     
     private final boolean isCheckingMetaData;
-    
+
     /**
      * Load table meta data.
-     * 
+     *
      * @param logicTableName logic table name
      * @param databaseType database type
      * @return table meta data
      * @throws SQLException SQL exception
      */
     public TableMetaData load(final String logicTableName, final DatabaseType databaseType) throws SQLException {
+        return load(logicTableName, databaseType, null);
+    }
+    
+    /**
+     * Load table meta data.
+     * 
+     * @param logicTableName logic table name
+     * @param databaseType database type
+     * @param schemaMetaData origin schema metadata
+     * @return table meta data
+     * @throws SQLException SQL exception
+     */
+    public TableMetaData load(final String logicTableName, final DatabaseType databaseType, final SchemaMetaData schemaMetaData) throws SQLException {
         TableRule tableRule = shardingRule.getTableRule(logicTableName);
         if (!isCheckingMetaData) {
             DataNode dataNode = tableRule.getActualDataNodes().iterator().next();
-            return TableMetaDataLoader.load(dataSourceMap.get(shardingRule.getShardingDataSourceNames().getRawMasterDataSourceName(
-                dataNode.getDataSourceName())), dataNode.getTableName(), databaseType.getName());
+            return null != schemaMetaData ? schemaMetaData.get(dataNode.getTableName())
+                    : TableMetaDataLoader.load(dataSourceMap.get(shardingRule.getShardingDataSourceNames()
+                     .getRawMasterDataSourceName(dataNode.getDataSourceName())), dataNode.getTableName(), databaseType.getName());
         }
         Map<String, List<DataNode>> dataNodeGroups = tableRule.getDataNodeGroups();
         Map<String, TableMetaData> actualTableMetaDataMap = new HashMap<>(dataNodeGroups.size(), 1);
@@ -118,20 +132,21 @@ public final class ShardingMetaDataLoader {
      * Load schema Meta data.
      *
      * @param databaseType database type
+     * @param schemaMetaData origin schema Meta data
      * @return schema Meta data
      * @throws SQLException SQL exception
      */
-    public SchemaMetaData load(final DatabaseType databaseType) throws SQLException {
-        SchemaMetaData result = loadShardingSchemaMetaData(databaseType);
+    public SchemaMetaData load(final DatabaseType databaseType, final SchemaMetaData schemaMetaData) throws SQLException {
+        SchemaMetaData result = loadShardingSchemaMetaData(databaseType, schemaMetaData);
         result.merge(loadDefaultSchemaMetaData(databaseType));
         return result;
     }
     
-    private SchemaMetaData loadShardingSchemaMetaData(final DatabaseType databaseType) throws SQLException {
+    private SchemaMetaData loadShardingSchemaMetaData(final DatabaseType databaseType, final SchemaMetaData schemaMetaData) throws SQLException {
         log.info("Loading {} logic tables' meta data.", shardingRule.getTableRules().size());
         Map<String, TableMetaData> tableMetaDataMap = new HashMap<>(shardingRule.getTableRules().size(), 1);
         for (TableRule each : shardingRule.getTableRules()) {
-            tableMetaDataMap.put(each.getLogicTable(), load(each.getLogicTable(), databaseType));
+            tableMetaDataMap.put(each.getLogicTable(), load(each.getLogicTable(), databaseType, schemaMetaData));
         }
         return new SchemaMetaData(tableMetaDataMap);
     }
